Move `mem_profile_task` into `chia.util.profiler`
