+++ chia/wallet/wallet_node.py
@@ -49,8 +49,7 @@ from chia.util.errors import KeychainIsLocked, KeychainProxyConnectionFailure, K
 from chia.util.ints import uint32, uint64
 from chia.util.keychain import Keychain
 from chia.util.path import path_from_root
-from chia.util.profiler import profile_task
-from chia.util.memory_profiler import mem_profile_task
+from chia.util.profiler import mem_profile_task, profile_task
 from chia.wallet.transaction_record import TransactionRecord
 from chia.wallet.util.new_peak_queue import NewPeakItem, NewPeakQueue, NewPeakQueueTypes
 from chia.wallet.util.peer_request_cache import PeerRequestCache, can_use_peer_request_cache
